requirements.txt
google-ai-generativelanguage==0.4.0
google-generativeai==0.3.1
streamlit==1.29.0


